ai-python-showcase/requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
